+++ src/swagger/v1/model/CpuInfoResult.h
@@ -58,8 +58,8 @@ public:
     /// <summary>
     /// Processor cache line size
     /// </summary>
-    int64_t getCacheLineSize() const;
-    void setCacheLineSize(int64_t value);
+    int32_t getCacheLineSize() const;
+    void setCacheLineSize(int32_t value);
     bool cacheLineSizeIsSet() const;
     void unsetCache_line_size();
     /// <summary>
@@ -73,7 +73,7 @@ public:
 protected:
     int32_t m_Cores;
     bool m_CoresIsSet;
-    int64_t m_Cache_line_size;
+    int32_t m_Cache_line_size;
     bool m_Cache_line_sizeIsSet;
     std::string m_Architecture;
     bool m_ArchitectureIsSet;
